Add files via upload
Worked with type in your XD wireframes and with InDesign
in Adobe XD. Created quality mockups and prototypes for
different screen resolutions. Designed and developed a Project 'Day-Care'. Parents nowadays are busy working to meet the family’s
needs. The problem came when they do not have much time
to take care of their children. Daycare is an application that
makes it easy for parents to find, book a daycare and also
monitor what activities are carried out by the child.