Phone Website Project
This is a simple Javascript (Frontend) website Created by using Html, CSS, Javascript, and JQuery. The purpose of this website is to provide Only Smartphone Like Android And Apple(ios) this Website is especially only For Phone Shopping. People can visit the website and after viewing all categories of smartphones they can easily Visualise which product is suitable for them and they don't need to go for shop anywhere else easily shopping facility is provided through the website they need to order the product and home delivery facility is there it also save a lot of time of the customer.